requirements.txt
Flask==2.0.2
numpy==1.22.4
pandas==1.3.5
gunicorn==20.1.0
Pillow==9.3
matplotlib==3.5.1
tensorflow==2.11.0
h5py>=2.9.0
gevent==20.9.0
